Issue 45 error annotation

+++ moead_framework/problem/combinatorial/mubqp.py
@@ -26,7 +26,7 @@ class Mubqp(Problem):
     >>> print(solution.F)
     """
 
-    dtype = int
+    dtype = float
 
     def __init__(self, instance_file):
         """
